Some fixes and updates for conversion microservice

+++ src/main/java/com/studyforces/sourcesapi/models/UploadConvertedFile.java
@@ -8,5 +8,5 @@ import javax.persistence.Embeddable;
 @Embeddable
 public class UploadConvertedFile {
     private String file;
-    private Long page;
+    private Integer page;
 }
